server/requirements.txt
# packages in environment at /Users/zoobaziz/anaconda3/envs/ChatServer:
#
# Name                    Version                   Build  Channel
accelerate                0.20.3                   pypi_0    pypi
aiohttp                   3.8.5                    pypi_0    pypi
aiosignal                 1.3.1                    pypi_0    pypi
asgiref                   3.7.2                    pypi_0    pypi
async-timeout             4.0.3                    pypi_0    pypi
attrs                     23.1.0                   pypi_0    pypi
autobahn                  23.6.2                   pypi_0    pypi
automat                   22.10.0                  pypi_0    pypi
backoff                   2.2.1                    pypi_0    pypi
bitsandbytes              0.41.1                   pypi_0    pypi
bzip2                     1.0.8                h620ffc9_4  
ca-certificates           2023.05.30           hca03da5_0  
certifi                   2023.7.22                pypi_0    pypi
cffi                      1.15.1                   pypi_0    pypi
channels                  4.0.0                    pypi_0    pypi
charset-normalizer        3.2.0                    pypi_0    pypi
click                     8.1.6                    pypi_0    pypi
constantly                15.1.0                   pypi_0    pypi
cryptography              41.0.3                   pypi_0    pypi
daphne                    4.0.0                    pypi_0    pypi
deprecated                1.2.14                   pypi_0    pypi
diskcache                 5.6.1                    pypi_0    pypi
django                    4.2.4                    pypi_0    pypi
einops                    0.6.1                    pypi_0    pypi
fairscale                 0.4.13                   pypi_0    pypi
filelock                  3.12.2                   pypi_0    pypi
fire                      0.5.0                    pypi_0    pypi
frozenlist                1.4.0                    pypi_0    pypi
fsspec                    2023.6.0                 pypi_0    pypi
googleapis-common-protos  1.60.0                   pypi_0    pypi
grpc-interceptor          0.15.2                   pypi_0    pypi
grpcio                    1.57.0                   pypi_0    pypi
grpcio-reflection         1.57.0                   pypi_0    pypi
grpcio-status             1.57.0                   pypi_0    pypi
grpcio-tools              1.51.1                   pypi_0    pypi
hf-transfer               0.1.3                    pypi_0    pypi
huggingface-hub           0.14.1                   pypi_0    pypi
hyperlink                 21.0.0                   pypi_0    pypi
idna                      3.4                      pypi_0    pypi
incremental               22.10.0                  pypi_0    pypi
jinja2                    3.1.2                    pypi_0    pypi
libffi                    3.4.4                hca03da5_0  
llama-cpp-python          0.1.78                   pypi_0    pypi
loguru                    0.6.0                    pypi_0    pypi
markupsafe                2.1.3                    pypi_0    pypi
mpmath                    1.3.0                    pypi_0    pypi
multidict                 6.0.4                    pypi_0    pypi
mypy-protobuf             3.4.0                    pypi_0    pypi
ncurses                   6.4                  h313beb8_0  
networkx                  3.1                      pypi_0    pypi
numpy                     1.25.2                   pypi_0    pypi
openssl                   3.0.10               h1a28f6b_0  
opentelemetry-api         1.15.0                   pypi_0    pypi
opentelemetry-exporter-otlp 1.15.0                   pypi_0    pypi
opentelemetry-exporter-otlp-proto-grpc 1.15.0                   pypi_0    pypi
opentelemetry-exporter-otlp-proto-http 1.15.0                   pypi_0    pypi
opentelemetry-instrumentation 0.36b0                   pypi_0    pypi
opentelemetry-instrumentation-grpc 0.36b0                   pypi_0    pypi
opentelemetry-proto       1.15.0                   pypi_0    pypi
opentelemetry-sdk         1.15.0                   pypi_0    pypi
opentelemetry-semantic-conventions 0.36b0                   pypi_0    pypi
packaging                 23.1                     pypi_0    pypi
peft                      0.4.0                    pypi_0    pypi
pillow                    10.0.0                   pypi_0    pypi
pip                       23.2.1          py311hca03da5_0  
protobuf                  4.24.0                   pypi_0    pypi
psutil                    5.9.5                    pypi_0    pypi
pyasn1                    0.5.0                    pypi_0    pypi
pyasn1-modules            0.3.0                    pypi_0    pypi
pycparser                 2.21                     pypi_0    pypi
pydantic                  1.10.12                  pypi_0    pypi
pyopenssl                 23.2.0                   pypi_0    pypi
python                    3.11.4               hb885b13_0  
pyyaml                    6.0.1                    pypi_0    pypi
readline                  8.2                  h1a28f6b_0  
regex                     2023.8.8                 pypi_0    pypi
requests                  2.31.0                   pypi_0    pypi
safetensors               0.3.2                    pypi_0    pypi
scipy                     1.11.1                   pypi_0    pypi
sentencepiece             0.1.99                   pypi_0    pypi
service-identity          23.1.0                   pypi_0    pypi
setuptools                68.0.0          py311hca03da5_0  
six                       1.16.0                   pypi_0    pypi
sqlite                    3.41.2               h80987f9_0  
sqlparse                  0.4.4                    pypi_0    pypi
sympy                     1.12                     pypi_0    pypi
termcolor                 2.3.0                    pypi_0    pypi
text-generation           0.6.0                    pypi_0    pypi
text-generation-server    1.0.1                    pypi_0    pypi
tk                        8.6.12               hb8d0fd4_0  
tokenizers                0.13.3                   pypi_0    pypi
torch                     2.0.1                    pypi_0    pypi
tqdm                      4.66.1                   pypi_0    pypi
transformers              4.31.0                   pypi_0    pypi
twisted                   22.10.0                  pypi_0    pypi
txaio                     23.1.1                   pypi_0    pypi
typer                     0.6.1                    pypi_0    pypi
types-protobuf            4.24.0.1                 pypi_0    pypi
typing-extensions         4.7.1                    pypi_0    pypi
tzdata                    2023c                h04d1e81_0  
urllib3                   2.0.4                    pypi_0    pypi
wheel                     0.38.4          py311hca03da5_0  
wrapt                     1.15.0                   pypi_0    pypi
xz                        5.4.2                h80987f9_0  
yarl                      1.9.2                    pypi_0    pypi
zlib                      1.2.13               h5a0b063_0  
zope-interface            6.0                      pypi_0    pypi
